Refactor tactical combat architecture
