+++ crates/adventuresim-core/src/autoresolve.rs
@@ -419,27 +419,22 @@ impl Combatant {
         }
     }
 
-    fn precision_damage_multiplier_cap_against(&self, defender: &Self) -> f32 {
-        let fallback = [BestiaryCategory::Human];
-        let categories = if defender.bestiary_categories.is_empty() {
-            &fallback
-        } else {
-            defender.bestiary_categories.as_slice()
-        };
-        let check = categories
-            .iter()
-            .map(|category| {
-                crate::capability::bestiary_knowledge_check(
-                    self.skills.bestiary_hours.effective(*category),
-                    self.attributes.instinct,
-                    self.attributes.intelligence,
-                    self.essentials.focus_level,
-                    self.body.body_part_health(BodyPart::Head),
-                )
-            })
-            .sum::<f32>()
-            / categories.len() as f32;
-        2.0 + check.clamp(0.0, 5.0)
+    fn view_with_equipment<'a>(
+        &'a self,
+        equipment: &'a CombatEquipment,
+    ) -> PlayerInfo<
+        &'a CombatAttributes,
+        &'a CombatBody,
+        &'a CombatEssentials,
+        &'a CombatEquipment,
+        &'a CombatSkills,
+    > {
+        PlayerInfo::empty()
+            .with_attributes(&self.attributes)
+            .with_body(&self.body)
+            .with_essentials(&self.essentials)
+            .with_equipment(equipment)
+            .with_skills(&self.skills)
     }
 
     pub fn incapacitation(&self) -> f32 {
@@ -1579,23 +1574,16 @@ fn melee_exchange(
     response: DefenderResponse,
 ) -> AttackResult {
     let attacker_equipment = attacker.equipment.for_melee();
-    resolve_melee_attack_by_parts(
-        &attacker.skills,
-        &attacker.attributes,
-        &attacker.body,
-        &attacker.essentials,
-        &attacker_equipment,
+    let attacker_view = attacker.view_with_equipment(&attacker_equipment);
+    let defender_view = defender.view_with_equipment(&defender.equipment);
+    attacker_view.resolve_melee_attack(
         attacker.equipment.holding_side,
+        &defender_view,
+        &defender.bestiary_categories,
+        response,
         precision,
-        attacker.precision_damage_multiplier_cap_against(defender),
         flanking,
         part,
-        response,
-        &defender.skills,
-        &defender.attributes,
-        &defender.body,
-        &defender.essentials,
-        &defender.equipment,
     )
 }
 
@@ -1608,22 +1596,15 @@ fn ranged_exchange(
     response: DefenderResponse,
 ) -> AttackResult {
     let attacker_equipment = attacker.equipment.for_ranged();
-    resolve_ranged_attack_by_parts(
-        &attacker.skills,
-        &attacker.attributes,
-        &attacker.body,
-        &attacker.essentials,
-        &attacker_equipment,
+    let attacker_view = attacker.view_with_equipment(&attacker_equipment);
+    let defender_view = defender.view_with_equipment(&defender.equipment);
+    attacker_view.resolve_ranged_attack(
+        &defender_view,
+        &defender.bestiary_categories,
+        response,
         precision,
-        attacker.precision_damage_multiplier_cap_against(defender),
         flanking,
         part,
-        response,
-        &defender.skills,
-        &defender.attributes,
-        &defender.body,
-        &defender.essentials,
-        &defender.equipment,
     )
 }
 
@@ -1914,10 +1895,14 @@ mod tests {
         attacker.skills.bestiary_hours.human = adventuresim_world_schema::BESTIARY_MASTERY_HOURS;
         let mut human = fighter(2, 1.0, false);
         human.bestiary_categories = vec![BestiaryCategory::Human];
-        let human_cap = attacker.precision_damage_multiplier_cap_against(&human);
+        let human_cap = attacker
+            .view_with_equipment(&attacker.equipment)
+            .precision_damage_multiplier_cap(&human.bestiary_categories);
 
         human.bestiary_categories = vec![BestiaryCategory::Human, BestiaryCategory::Draconid];
-        let combined_cap = attacker.precision_damage_multiplier_cap_against(&human);
+        let combined_cap = attacker
+            .view_with_equipment(&attacker.equipment)
+            .precision_damage_multiplier_cap(&human.bestiary_categories);
 
         assert!(human_cap > 2.0);
         assert!(combined_cap > 2.0);
