+++ wiki/reference/architecture.md
@@ -286,6 +286,11 @@ protocol. The tactical server validates melee allegiance, state, range, a fresh
 observed windup and cooldown, then authoritative physics line of sight before resolving an
 attack. Finite client-reported precision remains trusted because reconstructing
 animation and secondary physics is intentionally outside the headless server.
+The ordered wire messages are payload enums rather than phase-tagged field
+bags: starts cannot carry completion data, melee completions always name a
+target, and ranged completion distinguishes a miss from a targeted hit. Raw
+finite precision becomes `ReportedPrecision` without clamping or geometric
+reconstruction, and duration-backed authority types gate mutation.
 Accepted results mutate replicated limb health plus transient blood loss and
 imbalance. Shared autoresolve rules derive pain, blood-loss, and imbalance
 incapacitation and recover balance over time. Tactical enrollment projects
@@ -294,6 +299,9 @@ contributions; the same shared derivation as autoresolve excludes pain and blood
 from starting incapacitation before recomputing them live. Actors currently
 over the threshold stop moving, attacking, defending, and participating in
 offensive AI target selection; imbalance-only incapacitation can recover.
+The numeric incapacitation value is the sole stored readiness authority;
+active, staggered, and incapacitated status are mechanically derived from it
+rather than synchronized through a second boolean or ECS marker.
 
 These per-tick effects remain in memory only. A mission enemy's first transition
 into incapacitation counts as its defeat; recovery and later incapacitation do
@@ -315,6 +323,10 @@ Victory/Defeat presentation event, keeps the transport alive for a bounded
 three-second display window, and then exits. The delay is strictly post-commit:
 it cannot defer strategic authority or create a second outcome. A configured
 timeout remains a bounded `Failed` fallback.
+Enrollment and terminal progress are private lifecycle enums. A resolution and
+its bounded receipt form one frozen value whose retry time, acknowledgement
+deadline, transport failure, committed presentation, and finished state exist
+only in their applicable variants.
 
 The terminal call carries a bounded authenticated consequence receipt frozen
 with the resolution. It contains only Party character IDs, applied (clamped)
